backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-dateutil==2.9.0.post0
pydantic==2.9.2
psycopg2-binary==2.9.9
